requirements.txt
django==1.10.4
django-guardian==1.4.6
djangorestframework-filters==0.9.1
djangorestframework==3.5.3
markdown==2.6.7
django-filter==0.15.3
djangorestframework-jwt==1.9.0
mysqlclient==1.3.9
bcrypt==3.1.1
django-extensions==1.7.5
requests==2.12.4
django-debug-toolbar==1.6
#django-compressor==2.1
#csscompressor==0.9.4
django-redis==4.6
flake8==3.2.1
ipython==5.1.0
